extracted ability_spec from user_spec
